x86: IDT vector realignment (XCC)
Lays out what each IDT vector is used for, somewhat merging with the
Plan 9 stuff.  These are the vectors that the processor/LAPIC sees, not
the external IRQs.

This also changes the trap syscall vector to 0x40 (64), and the LAPIC
timer vector.

You need to reinstall your kernel headers and rebuild anything that made
trap based syscalls (parlib and all MCPs).  Not sure if you need to
rebuild the toolchain (probably not).