x86: one register_irq() for all types of buses
The 'raw' vectors are just the BusLAPIC type.  I didn't like the nxm way
of relying on BUSUNKNOWN to detect if it was potentially LAPIC.  That
meant you'd always need a bus (like BusPCI) and not just the IRQ for any
external IRQ numbers that were within the LAPIC vector range.

Another way to put it: the LAPIC 'irq' and the global system irq
(devices/acpi) share a different domain, so we need to differentiate on
the bus.